убрал таймслип, переименовал локаторы, исправил driver.quit()
